Refractor DeleteItemCommand to be abstract and add DeleteItemCommandByIndex
