+++ src/main/java/seedu/restaurant/logic/commands/menu/DeleteItemCommand.java
@@ -1,24 +1,15 @@
 package seedu.restaurant.logic.commands.menu;
 
-import static java.util.Objects.requireNonNull;
-
-import java.util.List;
-
-import seedu.restaurant.commons.core.EventsCenter;
-import seedu.restaurant.commons.core.Messages;
-import seedu.restaurant.commons.core.index.Index;
-import seedu.restaurant.commons.events.ui.DisplayItemListRequestEvent;
 import seedu.restaurant.logic.CommandHistory;
 import seedu.restaurant.logic.commands.Command;
 import seedu.restaurant.logic.commands.CommandResult;
 import seedu.restaurant.logic.commands.exceptions.CommandException;
 import seedu.restaurant.model.Model;
-import seedu.restaurant.model.menu.Item;
 
 /**
  * Deletes an item identified using it's displayed index from the menu.
  */
-public class DeleteItemCommand extends Command {
+public abstract class DeleteItemCommand extends Command {
 
     public static final String COMMAND_WORD = "delete-item";
 
@@ -31,42 +22,9 @@ public class DeleteItemCommand extends Command {
 
     public static final String MESSAGE_DELETE_ITEM_SUCCESS = "Deleted %1$d items";
 
-    private final Index targetIndex;
-    private final Index endingIndex;
-
-    public DeleteItemCommand(Index targetIndex, Index endingIndex) {
-        requireNonNull(targetIndex);
-
-        this.targetIndex = targetIndex;
-        this.endingIndex = endingIndex;
-    }
+    @Override
+    public abstract CommandResult execute(Model model, CommandHistory history) throws CommandException;
 
     @Override
-    public CommandResult execute(Model model, CommandHistory history) throws CommandException {
-        requireNonNull(model);
-        List<Item> lastShownList = model.getFilteredItemList();
-
-        if (endingIndex.getZeroBased() >= lastShownList.size()) {
-            throw new CommandException(Messages.MESSAGE_INVALID_ITEM_DISPLAYED_INDEX);
-        }
-
-        int numOfItems = endingIndex.getOneBased() - targetIndex.getZeroBased();
-
-        for (int i = endingIndex.getZeroBased(); i >= targetIndex.getZeroBased(); i--) {
-            Item itemToDelete = lastShownList.get(i);
-            model.deleteItem(itemToDelete);
-        }
-
-        model.commitRestaurantBook();
-        EventsCenter.getInstance().post(new DisplayItemListRequestEvent());
-        return new CommandResult(String.format(MESSAGE_DELETE_ITEM_SUCCESS, numOfItems));
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        return other == this // short circuit if same object
-                || (other instanceof DeleteItemCommand // instanceof handles nulls
-                    && targetIndex.equals(((DeleteItemCommand) other).targetIndex)
-                    && endingIndex.equals(((DeleteItemCommand) other).endingIndex)); // state check
-    }
+    public abstract boolean equals(Object other);
 }
