+++ src/test/java/seedu/restaurant/logic/commands/menu/DeleteItemCommandByIndexTest.java
@@ -27,7 +27,7 @@ import seedu.restaurant.model.menu.Item;
  * Contains integration tests (interaction with the Model, UndoCommand and RedoCommand) and unit tests for {@code
  * DeleteItemCommand}.
  */
-public class DeleteItemCommandTest {
+public class DeleteItemCommandByIndexTest {
 
     private Model model = new ModelManager(getTypicalRestaurantBook(), new UserPrefs());
     private CommandHistory commandHistory = new CommandHistory();
@@ -35,7 +35,7 @@ public class DeleteItemCommandTest {
     @Test
     public void execute_validIndexUnfilteredList_success() {
         Item itemToDelete = model.getFilteredItemList().get(INDEX_FIRST.getZeroBased());
-        DeleteItemCommand deleteItemCommand = new DeleteItemCommand(INDEX_FIRST, INDEX_FIRST);
+        DeleteItemCommand deleteItemCommand = new DeleteItemCommandByIndex(INDEX_FIRST, INDEX_FIRST);
 
         String expectedMessage = String.format(DeleteItemCommand.MESSAGE_DELETE_ITEM_SUCCESS, 1);
 
@@ -49,14 +49,14 @@ public class DeleteItemCommandTest {
     @Test
     public void execute_invalidIndexUnfilteredList_throwsCommandException() {
         Index outOfBoundIndex = Index.fromOneBased(model.getFilteredItemList().size() + 1);
-        DeleteItemCommand deleteItemCommand = new DeleteItemCommand(outOfBoundIndex, outOfBoundIndex);
+        DeleteItemCommand deleteItemCommand = new DeleteItemCommandByIndex(outOfBoundIndex, outOfBoundIndex);
 
         assertCommandFailure(deleteItemCommand, model, commandHistory, Messages.MESSAGE_INVALID_ITEM_DISPLAYED_INDEX);
     }
 
     @Test
     public void execute_validIndexUnfilteredListMultiple_success() {
-        DeleteItemCommand deleteItemCommand = new DeleteItemCommand(INDEX_FIRST, INDEX_THIRD);
+        DeleteItemCommand deleteItemCommand = new DeleteItemCommandByIndex(INDEX_FIRST, INDEX_THIRD);
         String expectedMessage = String.format(DeleteItemCommand.MESSAGE_DELETE_ITEM_SUCCESS, 3);
 
         ModelManager expectedModel = new ModelManager(model.getRestaurantBook(), new UserPrefs());
@@ -75,7 +75,7 @@ public class DeleteItemCommandTest {
         showItemAtIndex(model, INDEX_FIRST);
 
         Item itemToDelete = model.getFilteredItemList().get(INDEX_FIRST.getZeroBased());
-        DeleteItemCommand deleteItemCommand = new DeleteItemCommand(INDEX_FIRST, INDEX_FIRST);
+        DeleteItemCommand deleteItemCommand = new DeleteItemCommandByIndex(INDEX_FIRST, INDEX_FIRST);
 
         String expectedMessage = String.format(DeleteItemCommand.MESSAGE_DELETE_ITEM_SUCCESS, 1);
 
@@ -95,7 +95,7 @@ public class DeleteItemCommandTest {
         // ensures that outOfBoundIndex is still in bounds of restaurant book list
         assertTrue(outOfBoundIndex.getZeroBased() < model.getRestaurantBook().getItemList().size());
 
-        DeleteItemCommand deleteItemCommand = new DeleteItemCommand(outOfBoundIndex, outOfBoundIndex);
+        DeleteItemCommand deleteItemCommand = new DeleteItemCommandByIndex(outOfBoundIndex, outOfBoundIndex);
 
         assertCommandFailure(deleteItemCommand, model, commandHistory, Messages.MESSAGE_INVALID_ITEM_DISPLAYED_INDEX);
     }
@@ -103,7 +103,7 @@ public class DeleteItemCommandTest {
     @Test
     public void executeUndoRedo_validIndexUnfilteredList_success() throws Exception {
         Item itemToDelete = model.getFilteredItemList().get(INDEX_FIRST.getZeroBased());
-        DeleteItemCommand deleteItemCommand = new DeleteItemCommand(INDEX_FIRST, INDEX_FIRST);
+        DeleteItemCommand deleteItemCommand = new DeleteItemCommandByIndex(INDEX_FIRST, INDEX_FIRST);
         Model expectedModel = new ModelManager(model.getRestaurantBook(), new UserPrefs());
         expectedModel.deleteItem(itemToDelete);
         expectedModel.commitRestaurantBook();
@@ -123,7 +123,7 @@ public class DeleteItemCommandTest {
     @Test
     public void executeUndoRedo_invalidIndexUnfilteredList_failure() {
         Index outOfBoundIndex = Index.fromOneBased(model.getFilteredItemList().size() + 1);
-        DeleteItemCommand deleteItemCommand = new DeleteItemCommand(outOfBoundIndex, outOfBoundIndex);
+        DeleteItemCommand deleteItemCommand = new DeleteItemCommandByIndex(outOfBoundIndex, outOfBoundIndex);
 
         // execution failed -> restaurant book state not added into model
         assertCommandFailure(deleteItemCommand, model, commandHistory, Messages.MESSAGE_INVALID_ITEM_DISPLAYED_INDEX);
@@ -141,7 +141,7 @@ public class DeleteItemCommandTest {
      */
     @Test
     public void executeUndoRedo_validIndexFilteredList_sameItemDeleted() throws Exception {
-        DeleteItemCommand deleteItemCommand = new DeleteItemCommand(INDEX_FIRST, INDEX_FIRST);
+        DeleteItemCommand deleteItemCommand = new DeleteItemCommandByIndex(INDEX_FIRST, INDEX_FIRST);
         Model expectedModel = new ModelManager(model.getRestaurantBook(), new UserPrefs());
 
         showItemAtIndex(model, INDEX_SECOND);
@@ -164,14 +164,14 @@ public class DeleteItemCommandTest {
 
     @Test
     public void equals() {
-        DeleteItemCommand deleteFirstCommand = new DeleteItemCommand(INDEX_FIRST, INDEX_FIRST);
-        DeleteItemCommand deleteSecondCommand = new DeleteItemCommand(INDEX_SECOND, INDEX_SECOND);
+        DeleteItemCommand deleteFirstCommand = new DeleteItemCommandByIndex(INDEX_FIRST, INDEX_FIRST);
+        DeleteItemCommand deleteSecondCommand = new DeleteItemCommandByIndex(INDEX_SECOND, INDEX_SECOND);
 
         // same object -> returns true
         assertTrue(deleteFirstCommand.equals(deleteFirstCommand));
 
         // same values -> returns true
-        DeleteItemCommand deleteFirstCommandCopy = new DeleteItemCommand(INDEX_FIRST, INDEX_FIRST);
+        DeleteItemCommand deleteFirstCommandCopy = new DeleteItemCommandByIndex(INDEX_FIRST, INDEX_FIRST);
         assertTrue(deleteFirstCommand.equals(deleteFirstCommandCopy));
 
         // different types -> returns false
